+++ R/reproj.R
@@ -34,7 +34,7 @@
 #' to `target` is applied (this can be controlled by setting options).
 #'
 #' The function [reproj()] always returns a 3-column matrix _unless_ `four =
-#' TRUE`, and [PROJ::ok_proj6()] is `TRUE` and then a 4-column matrix is returned.
+#' TRUE`, and PROJ package is available then a 4-column matrix is returned.
 #'
 #' Functions [reproj_xy()] and [reproj_xyz()] are helpers for [reproj()] and always
 #' return 2- or 3-column matrix respectively.
@@ -114,8 +114,15 @@ reproj <- function(x, target, ..., source = NULL, four = FALSE) {
 
 
 #' @rdname reproj
+#' @importFrom utils packageVersion
 #' @export
 reproj.matrix <- function(x, target, ..., source = NULL, four = FALSE) {
+  if (packageVersion("PROJ") <= "0.4.5") {
+    x <- x[,1:2, drop = FALSE]
+  }
+  nms <- colnames(x)
+  ## make sure all columns have names, or none
+  if (any(nzchar(nms))) colnames(x) <- NULL
   if (isTRUE(four)) {
     stop("argument 'four' is not available currently")
   }
@@ -127,56 +134,18 @@ reproj.matrix <- function(x, target, ..., source = NULL, four = FALSE) {
     } else {
       stop("no 'source' projection included, and does not look like longitude/latitude values")
     }
-  } else {
-    source <- to_proj(source)
   }
-  target <- to_proj(target)  ## just sprintf("EPSG:%i", target) or sprintf("+init=epsg:%i", target)
-  if (PROJ::ok_proj6()) {
 
-    if (dim(x)[2L] == 2L) {
+  if (.ok_PROJ()) {
+
+   
       out <- PROJ::proj_trans(x, target = target, ..., source = source)
-      out <- cbind(do.call(cbind, out), 0)
-
-    }
-    if (dim(x)[2L] == 3L) {
-      out <- PROJ::proj_trans(x[,1:2, drop = FALSE], target = target, ..., source = source,
-                                      z_ = x[, 3L, drop = TRUE])
-      out <- do.call(cbind, out)
-      if (!four) out <- out[ , 1:3, drop = FALSE]
-    }
-    if (dim(x)[2L] > 3L) {
-      out <- PROJ::proj_trans(x[,1:2, drop = FALSE], target = target, ..., source = source,
-                                      z_ = x[, 3L, drop = TRUE],
-                                      t_ = x[, 4L, drop = TRUE])
-      out <- do.call(cbind, out)
-      if (!four) out <- out[ , 1:3, drop = FALSE]
-    }
+    
 
 
-    if (four) {
-      if (dim(out)[2] == 2) {
-        out <- cbind(out, 0, 0)
-      }
-      if (dim(out)[2] == 3) {
-        out <- cbind(out, 0)
-      }
-    }
-  } else {
-
-    target <- to_proj(target)
-    validate_proj(target)
-
-
-    source <- to_proj(source)
-    validate_proj(source)
-
-    srcmult <- if (is_ll(source)) {pi/180} else {1}
-    tarmult <-  if(is_ll(target)) {180/pi} else {1}
-
-    x[, 1:2] <- x[,1:2] * srcmult
-    out <- proj4::ptransform(x, source, target, ...)
-    out[,1:2] <- out[, 1:2] * tarmult
-    if (four) warning("argument 'four' is ignored when PROJ version 6 not available")
+  } 
+  if (packageVersion("PROJ") <= "0.4.5") {
+    out <- do.call(cbind, out)
   }
   out
 }
@@ -191,13 +160,15 @@ reproj.data.frame <- function(x, target, ..., source = NULL, four = FALSE) {
 #' @rdname reproj
 #' @export
 reproj_xy <- function(x, target, ..., source = NULL) {
-  reproj(x, target = target, source = source, ...)[,1:2]
+  reproj(x, target = target, source = source, ...)[,1:2, drop = FALSE]
 }
 
 #' @rdname reproj
 #' @export
 reproj_xyz <- function(x, target, ..., source = NULL) {
-  reproj(x, target = target, source = source, ...)[,1:3]
+  if (ncol(x) > 3L) x <- x[,1:3, drop = FALSE]
+  if (ncol(x) == 2L) x <- cbind(x, 0.0)
+  reproj(x, target = target, source = source, ...)[,1:3, drop = FALSE]
 }