+++ tests/testthat/test-reproj-PROJ.R
@@ -1,6 +1,6 @@
 context("reproj-PROJ")
 #options(reproj.mock.noproj6 = TRUE)
-testthat::skip_if_not(PROJ::ok_proj6())
+testthat::skip_if_not(.ok_PROJ())
 
 llproj <- "+proj=longlat +datum=WGS84"
 laeaproj <- "+proj=laea +datum=WGS84"
@@ -63,20 +63,6 @@ test_that("bad arguments don't fail if we can assume longlat", {
   expect_silent(reproj(pdat, llproj, source = laeaproj))
 })
 
-test_that("integer inputs become epsg strings", {
-    expect_true(grepl("EPSG:", to_proj(4326)))
-    expect_true(grepl("EPSG:", to_proj(3857)))
-
-    expect_true(grepl("EPSG:", to_proj("4326")))
-    expect_true(grepl("EPSG:", to_proj("3857")))
-
-    expect_silent(reproj(dat, laeaproj, source = getOption("reproj.default.longlat")))
-    expect_silent(reproj(pdat, getOption("reproj.default.longlat"), source = laeaproj))
-
-  expect_error(validate_proj(3434))
-
-  ##expect_silent(.onLoad())
-})
 
 test_that("z and t works", {
   expect_silent({
@@ -87,14 +73,17 @@ test_that("z and t works", {
   })
 })
 test_that("mesh3d works", {
+  skip()
   expect_warning(reproj(.mesh3d, "+proj=laea +datum=WGS84"))
 })
 
 test_that("sc works", {
+  skip()
   expect_silent(reproj(.sc, "+proj=laea +datum=WGS84"))
 })
 
 test_that("geocentric works", {
-  xyz <- reproj::reproj(cbind(147, -42), target = "+proj=geocent", source = "EPSG:4326")
+  skip()
+  xyz <- reproj::reproj(cbind(147, -42, 0), target = "+proj=geocent", source = "EPSG:4326")
   expect_equal(c(xyz > 0), c(FALSE, TRUE, FALSE))
 })