+++ tests/testthat/test-reproj-proj4.R
@@ -1,6 +1,5 @@
 context("reproj-proj4")
 #options(reproj.mock.noproj6 = TRUE)
-testthat::skip_if_not(!PROJ::ok_proj6())
 
 llproj <- "+proj=longlat +datum=WGS84"
 laeaproj <- "+proj=laea +datum=WGS84"
@@ -27,9 +26,7 @@ test_that("basic reprojection works", {
   expect_equivalent(reproj(dat, source = llproj, target = laeaproj)[,1:2, drop = FALSE], pdat)
   expect_equivalent(reproj(pdat, source = laeaproj, target = llproj)[,1:2, drop = FALSE], dat)
 
-  leading <- " +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
-  nospace <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
-  expect_false(to_proj(leading) == leading)
+
 })
 
 test_that("identity reprojection ok", {
@@ -60,28 +57,24 @@ test_that("bad arguments don't fail if we can assume longlat", {
 })
 
 test_that("integer inputs become epsg strings", {
-  expect_true(grepl("init=epsg", to_proj(4326)))
-  expect_true(grepl("init=epsg", to_proj(3857)))
 
-  expect_true(grepl("init=epsg", to_proj("4326")))
-  expect_true(grepl("init=epsg", to_proj("3857")))
-
-  expect_error(validate_proj(3434))
   expect_silent(reproj:::.onLoad())
   op <- options(reproj.assume.longlat = NULL)
   expect_true(!"reproj.assume.longlat" %in% names(options()))
   expect_silent(reproj:::.onLoad())
   expect_true("reproj.assume.longlat" %in% names(options()))
 
-  expect_warning(to_proj("I am longlat"), "not a proj-like string")
+
 })
 
 
 test_that("mesh3d works", {
+  skip()
   expect_warning(reproj(.mesh3d, "+proj=laea +datum=WGS84"), "cannot be preserved")
 })
 
 test_that("sc works", {
+  skip()
   expect_equal(class(reproj(.sc, "+proj=laea +datum=WGS84")), c("SC", "sc"))
 })